Print score

+++ Copter/Copter.ino
@@ -221,17 +221,17 @@ void setup()  {
   srand(0);
   state = createGameState();
     createCave(state);
-    
+
     long lBtn1;
     long lBtn2;
     printStart();
-     
+
     do {
       lBtn1 = GPIOPinRead(BTN1Port, BTN1);
       lBtn2 = GPIOPinRead(BTN2Port, BTN2);
     } while (lBtn1 != BTN1 && lBtn2 != BTN2);
 }
-  
+
 void render() {
   // OrbitOledClear();
   OrbitOledMoveTo(0,0);
@@ -265,15 +265,15 @@ void loop()  {
 
   readSensorData(state);
   updateGameLogic(state, (2.0 - SPEED/10) / 20); // For now, assume 20 FPS
-  
+
   //If collision is detected (gameover)
   if (state->gameOver == 1)
   {
-    printGameOver(state->score);
+    printGameOver((long) state->score);
     delay(5000);
     setup();
   }
-  
+
   render();
   delay(10-SPEED);
 }
