+++ Copter/print.h
@@ -1,7 +1,7 @@
 #ifndef PRINT_H
 #define PRINT_H
 
-void printGameOver(int score);
+void printGameOver(long score);
 void printStart();
 
 #endif
